Restored working position pointer (with some caveats) to trackeditor.cpp.
Some optimisations in Composition, and a new method for relatively
quickly finding the extents of a bar given a time in it.

Restored working current-time tracker to notationcanvasview.cpp.

Various other bits of cleaning & tidying.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1212 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/notationhlayout.cpp
@@ -484,7 +484,7 @@ NotationHLayout::reconcileBars()
 
 	    BarDataList &list = i->second;
 
-	    if (list.size() > barNo) {
+	    if ((int)list.size() > barNo) {
 
 		reachedEnd = false;
 
@@ -498,7 +498,7 @@ NotationHLayout::reconcileBars()
 
 	    BarDataList &list = i->second;
 
-	    if (list.size() > barNo) {
+	    if ((int)list.size() > barNo) {
 
 		BarData &bd(list[barNo]);
